Addressed PR feedback.
