+++ libraries/Microsoft.Bot.Builder/ChannelServiceHandlerBase.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
 using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading;
