+++ libraries/Microsoft.Bot.Builder/Skills/CloudSkillHandler.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
