+++ libraries/Microsoft.Bot.Connector/Authentication/BuiltinBotFrameworkAuthentication.cs
@@ -67,17 +67,7 @@ namespace Microsoft.Bot.Connector.Authentication
             return _toChannelFromBotOAuthScope;
         }
 
-        public override Task<AppCredentials> GetAppCredentialsAsync(string appId, HttpClient client, string oAuthScope, CancellationToken cancellationToken)
-        {
-            // TODO: fix hack
-            var factory = _credentialFactory as PasswordServiceClientCredentialFactory;
-            AppCredentials credentials = _callerId == CallerIdConstants.USGovChannel
-                ? new MicrosoftGovernmentAppCredentials(appId, factory?.Password, client, NullLogger.Instance, oAuthScope)
-                : new MicrosoftAppCredentials(appId, factory?.Password, client, NullLogger.Instance, oAuthScope);
-            return Task.FromResult(credentials);
-        }
-
-        public override async Task<ClaimsIdentity> ValidateAuthHeaderAsync(string authHeader, bool isSkillCallback, CancellationToken cancellationToken)
+        public override async Task<ClaimsIdentity> ValidateSkillsAuthHeaderAsync(string authHeader, CancellationToken cancellationToken)
         {
             if (string.IsNullOrWhiteSpace(authHeader))
             {
@@ -91,9 +81,8 @@ namespace Microsoft.Bot.Connector.Authentication
                 // In the scenario where auth is disabled, we still want to have the
                 // IsAuthenticated flag set in the ClaimsIdentity.
                 // To do this requires adding in an empty claim.
-                return isSkillCallback
-                    ? SkillValidation.CreateAnonymousSkillClaim()
-                    : new ClaimsIdentity(new List<Claim>(), AuthenticationConstants.AnonymousAuthType);
+                // Since ChannelServiceHandler calls are always a skill callback call, we set the skill claim too.
+                return SkillValidation.CreateAnonymousSkillClaim();
             }
 
             return await JwtTokenValidation
