+++ libraries/integration/Microsoft.Bot.Builder.Integration.AspNet.Core/Skills/SkillHttpClient.cs
@@ -95,7 +95,8 @@ namespace Microsoft.Bot.Builder.Integration.AspNet.Core.Skills
         /// <returns>Async task with optional invokeResponse of type T.</returns>
         public virtual async Task<InvokeResponse<T>> PostActivityAsync<T>(string fromBotId, BotFrameworkSkill toSkill, Uri callbackUrl, Activity activity, CancellationToken cancellationToken)
         {
-            return await PostActivityAsync<T>(GetOriginatingAudience(), fromBotId, toSkill, callbackUrl, activity, cancellationToken).ConfigureAwait(false);
+            var originatingAudience = ChannelProvider != null && ChannelProvider.IsGovernment() ? GovernmentAuthenticationConstants.ToChannelFromBotOAuthScope : AuthenticationConstants.ToChannelFromBotOAuthScope;
+            return await PostActivityAsync<T>(originatingAudience, fromBotId, toSkill, callbackUrl, activity, cancellationToken).ConfigureAwait(false);
         }
     }
 }
